Fixed small error, standardized comparing parameters, and refactored
